splitting files and reorganising code
