+++ lib/src/ui/pages/home_page.dart
@@ -1,0 +1,101 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import '../../logic/my_app_state.dart';
+import '../../logic/my_favourites_page_state.dart';
+import '../elements/navigation_rail.dart';
+import '../elements/navigation_bar.dart';
+import 'generator_page.dart';
+import 'my_favourites_page.dart';
+
+class MyHomePage extends StatefulWidget {
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  var selectedIndex = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    Widget page;
+    switch (selectedIndex) {
+      case 0:
+        page = GeneratorPage();
+      case 1:
+        page = ChangeNotifierProvider(
+          create: (context) => MyFavouritesPageState(context.read<MyAppState>()),
+          child: MyFavouritesPage(),
+        );
+      default:
+        throw UnimplementedError('no widget for $selectedIndex');
+    }
+
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        if (constraints.maxWidth >= 1024) {
+          // Desktop
+          return Scaffold(
+            body: Row(
+              children: [
+                CustomNavigationRail(
+                  selectedIndex: selectedIndex,
+                  onDestinationSelected: (value) {
+                    setState(() {
+                      selectedIndex = value;
+                    });
+                  },
+                  extended: true,
+                ),
+                Expanded(
+                  child: Container(
+                    color: Theme.of(context).colorScheme.primaryContainer,
+                    child: page,
+                  ),
+                ),
+              ],
+            ),
+          );
+        } else if (constraints.maxWidth >= 600) {
+          // Tablet
+          return Scaffold(
+            body: Row(
+              children: [
+                CustomNavigationRail(
+                  selectedIndex: selectedIndex,
+                  onDestinationSelected: (value) {
+                    setState(() {
+                      selectedIndex = value;
+                    });
+                  },
+                  extended: false,
+                ),
+                Expanded(
+                  child: Container(
+                    color: Theme.of(context).colorScheme.primaryContainer,
+                    child: page,
+                  ),
+                ),
+              ],
+            ),
+          );
+        } else {
+          // Mobile
+          return Scaffold(
+            body: Container(
+              color: Theme.of(context).colorScheme.primaryContainer,
+              child: page,
+            ),
+            bottomNavigationBar: CustomNavigationBar(
+              selectedIndex: selectedIndex,
+              onDestinationSelected: (value) {
+                setState(() {
+                  selectedIndex = value;
+                });
+              },
+            ),
+          );
+        }
+      },
+    );
+  }
+}
